Move deref to root and drop utils module
