--- src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod deref;
-pub(crate) mod value;
